% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cspstrain-package.R
\docType{package}
\name{cspstrain-package}
\alias{cspstrain}
\alias{cspstrain-package}
\title{cspstrain: CSPS microdata training}
\description{
Training materials for learning how to work with Civil Service
    People Survey (CSPS) in R.
}
\author{
\strong{Maintainer}: Matt Dray \email{matt.dray@cabinetoffice.gov.uk}

}
\keyword{internal}
